ez_ai_agent_demo/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0" 
anthropic==0.42.0 ; python_version >= "3.10" and python_version < "4.0" 
anyio==4.7.0 ; python_version >= "3.10" and python_version < "4.0" 
beautifulsoup4==4.12.3 ; python_version >= "3.10" and python_version < "4.0" 
cachetools==5.5.0 ; python_version >= "3.10" and python_version < "4.0" 
certifi==2024.12.14 ; python_version >= "3.10" and python_version < "4.0" 
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "4.0" 
cohere==5.13.4 ; python_version >= "3.10" and python_version < "4.0" 
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" 
distro==1.9.0 ; python_version >= "3.10" and python_version < "4.0" 
exceptiongroup==1.2.2 ; python_version >= "3.10" and python_version < "3.11" 
faiss-cpu==1.9.0.post1 ; python_version >= "3.10" and python_version < "4.0" 
fake-useragent==2.0.3 ; python_version >= "3.10" and python_version < "4.0" 
fastavro==1.10.0 ; python_version >= "3.10" and python_version < "4.0" 
filelock==3.16.1 ; python_version >= "3.10" and python_version < "4.0" 
fsspec==2024.12.0 ; python_version >= "3.10" and python_version < "4.0" 
google-ai-generativelanguage==0.6.10 ; python_version >= "3.10" and python_version < "4.0" 
google-api-core==2.24.0 ; python_version >= "3.10" and python_version < "4.0" 
google-api-core[grpc]==2.24.0 ; python_version >= "3.10" and python_version < "4.0" 
google-api-python-client==2.156.0 ; python_version >= "3.10" and python_version < "4.0" 
google-auth-httplib2==0.2.0 ; python_version >= "3.10" and python_version < "4.0" 
google-auth==2.37.0 ; python_version >= "3.10" and python_version < "4.0" 
google-generativeai==0.8.3 ; python_version >= "3.10" and python_version < "4.0" 
googleapis-common-protos==1.66.0 ; python_version >= "3.10" and python_version < "4.0" 
groq==0.13.1 ; python_version >= "3.10" and python_version < "4.0" 
grpcio-status==1.67.1 ; python_version >= "3.10" and python_version < "4.0" 
grpcio==1.67.1 ; python_version >= "3.10" and python_version < "4.0" 
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0" 
halo==0.0.31 ; python_version >= "3.10" and python_version < "4.0" 
httpcore==1.0.7 ; python_version >= "3.10" and python_version < "4.0" 
httplib2==0.22.0 ; python_version >= "3.10" and python_version < "4.0" 
httpx-sse==0.4.0 ; python_version >= "3.10" and python_version < "4.0" 
httpx==0.27.2 ; python_version >= "3.10" and python_version < "4.0" 
huggingface-hub==0.27.0 ; python_version >= "3.10" and python_version < "4.0" 
idna==3.10 ; python_version >= "3.10" and python_version < "4.0" 
igraph==0.11.8 ; python_version >= "3.10" and python_version < "4.0" 
jiter==0.8.2 ; python_version >= "3.10" and python_version < "4.0" 
leidenalg==0.10.2 ; python_version >= "3.10" and python_version < "4.0" 
log-symbols==0.0.14 ; python_version >= "3.10" and python_version < "4.0" 
lxml==5.3.0 ; python_version >= "3.10" and python_version < "4.0" 
numpy==2.2.1 ; python_version >= "3.10" and python_version < "4.0" 
ollama==0.4.4 ; python_version >= "3.10" and python_version < "4.0" 
openai==1.58.1 ; python_version >= "3.10" and python_version < "4.0" 
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0" 
parameterized==0.9.0 ; python_version >= "3.10" and python_version < "4.0" 
pinecone-plugin-inference==3.1.0 ; python_version >= "3.10" and python_version < "4.0" 
pinecone-plugin-interface==0.0.7 ; python_version >= "3.10" and python_version < "4.0" 
pinecone==5.4.2 ; python_version >= "3.10" and python_version < "4.0" 
proto-plus==1.25.0 ; python_version >= "3.10" and python_version < "4.0" 
protobuf==5.29.2 ; python_version >= "3.10" and python_version < "4.0" 
pyasn1-modules==0.4.1 ; python_version >= "3.10" and python_version < "4.0" 
pyasn1==0.6.1 ; python_version >= "3.10" and python_version < "4.0" 
pydantic-core==2.27.2 ; python_version >= "3.10" and python_version < "4.0" 
pydantic==2.10.4 ; python_version >= "3.10" and python_version < "4.0" 
pyparsing==3.2.0 ; python_version >= "3.10" and python_version < "4.0" 
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0" 
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0" 
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0" 
regex==2024.11.6 ; python_version >= "3.10" and python_version < "4.0" 
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0" 
rsa==4.9 ; python_version >= "3.10" and python_version < "4" 
sentence-splitter==1.4 ; python_version >= "3.10" and python_version < "4.0" 
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0" 
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0" 
soupsieve==2.6 ; python_version >= "3.10" and python_version < "4.0" 
spinners==0.0.24 ; python_version >= "3.10" and python_version < "4.0" 
termcolor==2.5.0 ; python_version >= "3.10" and python_version < "4.0" 
texttable==1.7.0 ; python_version >= "3.10" and python_version < "4.0" 
tokenizers==0.21.0 ; python_version >= "3.10" and python_version < "4.0" 
tqdm==4.67.1 ; python_version >= "3.10" and python_version < "4.0" 
types-requests==2.32.0.20241016 ; python_version >= "3.10" and python_version < "4.0" 
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0" 
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0" 
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "4.0" 
browser-use==0.1.37
steel-sdk==0.1.0b8
langchain-google-genai== 2.0.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
Flask-SSE==1.0.0
yahoofinance==0.0.2
yarl==1.18.3
yfinance==0.2.51